+++ backend/koala/database/model/link_docs_to_entity.py
@@ -4,7 +4,6 @@ from typing import Any, List
 from immudb import ImmudbClient
 
 from .document import DocumentID
-from .tool import ToolID
 
 
 @dataclass(unsafe_hash=True)
