+++ backend/koala/database/model/tool.py
@@ -1,5 +1,5 @@
 from dataclasses import dataclass
-from typing import Any, List
+from typing import List
 
 from immudb import ImmudbClient
 
@@ -18,36 +18,21 @@ class Tool(ToolID):
     ) -> None:
         super().__init__(name=name, version_major=version_major, purpose=purpose, gmp_relevant=gmp_relevant)
         self._client = client
-        self._entity = DataBaseEntity(client=client)
+        self._entity = DataBaseEntity(
+            client=client,
+            entity=ToolID(
+                name=name,
+                version_major=version_major,
+                purpose=purpose,
+                gmp_relevant=gmp_relevant,
+            ),
+        )
 
     def add(self) -> None:
-        self._entity.insert(
-            ToolID(
-                name=self.name,
-                version_major=self.version_major,
-                purpose=self.purpose,
-                gmp_relevant=self.gmp_relevant,
-            )
-        )
+        self._entity.insert()
 
     def get_id(self) -> int:
-        resp = self._client.sqlQuery(
-            """
-                SELECT id FROM entity
-                WHERE name = @name
-                AND version_major = @version_major
-                AND purpose = @purpose
-                AND is_system = FALSE;
-                """,
-            params={
-                "name": self.name,
-                "version_major": self.version_major,
-                "purpose": self.purpose,
-            },
-        )
-        if len(resp) == 1:
-            return int(resp[0][0])
-        raise Exception("Tool not found")
+        return self._entity.get_id()
 
 
 def _in(entitylinker: tuple, entity: tuple) -> bool:
