+++ backend/pyproject.toml
@@ -10,7 +10,6 @@ python = "^3.10"
 dataclasses-json = "^0.5.7"
 dependency-injector = "^4.40.0"
 immudb-py = "^1.3.2"
-bandit = "^1.7.4"
 
 [tool.poetry.dev-dependencies]
 pytest = "7.1.2"
@@ -21,6 +20,7 @@ black = "22.6.0"
 isort = "5.10.1"
 bandit = "1.7.4"
 pylint = "2.14.4"
+pytest-md-report = "^0.3.0"
 
 [tool.mypy]
 files = "koala/**/*.py"
