+++ backend/tests/test_change.py
@@ -1,0 +1,11 @@
+import pytest
+
+
+@pytest.mark.skip(reason="Not Implemented")
+def test_user_approves_given_change():
+    pass
+
+
+@pytest.mark.skip(reason="Not Implemented")
+def test_user_refuses_given_change():
+    pass
